move wallet and miner interfaces into their respective packages
