+++ .gitignore
@@ -1,3 +1,4 @@
+hostdir
 release
 whitepaper.aux
 whitepaper.log
